backend/requirements.txt
Flask
flask-cors
elasticsearch
google-cloud-bigquery
db-dtypes
requests